Missing exception specifier for basic_utf8_string c'tor
The "Constructor taking an utf8 char literal" c'tor has two overloaded member functions which have different exception specifiers however, as they are templated, it wasn't something that was picked up by the compiler.  When compiling under VS 2017, the difference in the exception specifiers (evaluating to "noexcept (true)"  vs. "noexcept (false)" based on the allocator being nothrow copy constructable) results in a compliation error.

The same, more verbose, exception specifier was added to the second member c'tor so that the exception specification is the same for both c'tors.