Correct noexcept specification (with allocators)
The existing noexcept specification (when specifying an allocator) incorrectly evaluated the TINY_UTF8_NOEXCEPT value and the noexcept specification of the c'tor(s) of the allocator using a logical OR.  This meant that a basic_utf8_string could be marked as noexcept(true) incorrectly as the allocator could still throw even if TINY_UTF8_NOEXCEPT was set (as T || F -> T).  Changing these to evaluate with a logical AND (i.e. T && F -> F) means that the correct noexcept specification is set.

Unit tests with a throwable and non-throwable allocator (along with TINY_UTF8_NOEXCEPT) demonstrate the correctness.