Fix unit test build issue on MacOS
Building the unit tests on MacOS, using XCode, results in a compilation error when compiling code using std::equal.  This is due to the use of std::equal(it, it, it, it, predicate) which is a prototype that is only available in C++14 onwards.  For C++11, the correct prototype would be std::equal(it,it,it, predicate).  On Linux and Windows, the former works fine but this fails on MacOS.  This change corrects the unit tests so that the latter is used (which is equivalent).

This change also enforces the use of standards compliance C++11 as part of the CMake scripts.