Fix for std::enable_if substitution failure on VS 2017 (toolset v141)
When calling the constructor (taking an utf8 char literal), VS 2017 incorrect evaluates the boolean expression and enables both SSO and non-SSO variants leading to a compilation error as the c'tors have the same prototype.  This is compounded by the Travis CI using this toolset when compiling jobs for Windows.  This change fixes this issue and permits compilation under the VS 2017 toolset.