Refactor unit tests and build system (#53)
* Refactor unit tests and build system

The change to the unit testing refactors them so that they:
- provide an explicit pass/fail when executed as part of the CI system
- are not dependent on any external test files (input and output)
- are grouped into logical test blocks so that new tests can be added

The change to the build system refactors it so that:
- a modern CMake approach is taken for a C++ header-only library
- provides for installation of the header-only library
- adds support for building documentation when a top-level project

* Add build dependency scripts

Update the Travis CI build scripts to call specific build scripts for each supported environments.

* Make install-dep-linux/osx.sh scripts executable

Mark the dependency installation scripts as executable so that Travis CI doesn't reject them with a "permission denied" error during the installation stage.

* Prevent vcpkg bootstrap from terminating script

Make the install-dep-win32 script spawn the vcpkg bootstrap script in a separate command process to prevent it from triggering the premature exiting of the dependencies script.

* Fix silly typos in Travis build files

Correct silly mistakes in the various Travis CI build files (e.g. target typos, missing sudo, and path slashes under Windows) that prevent all the builds from building.

* Fixes build systems on MacOS and Windows

Correct an issue with paths for CMake dependencies under Windows and move to the latest version of XCode.

* Specify MSBuild Generator under Windows

The Travis CI system needs to use the msbuild process (via CMake) so the correct generator and build steps need to be used.

* Missing exception specifier for basic_utf8_string c'tor

The "Constructor taking an utf8 char literal" c'tor has two overloaded member functions which have different exception specifiers however, as they are templated, it wasn't something that was picked up by the compiler.  When compiling under VS 2017, the difference in the exception specifiers (evaluating to "noexcept (true)"  vs. "noexcept (false)" based on the allocator being nothrow copy constructable) results in a compliation error.

The same, more verbose, exception specifier was added to the second member c'tor so that the exception specification is the same for both c'tors.

* Remove duplicate c'tor

There are two templates for "basic_utf8_string" c'tor taking a string literal and an allocator (differing by exception specification which isn't great).  Remove the second c'tor which will clash with the first when the exception specifications match).

* Correct noexcept specification (with allocators)

The existing noexcept specification (when specifying an allocator) incorrectly evaluated the TINY_UTF8_NOEXCEPT value and the noexcept specification of the c'tor(s) of the allocator using a logical OR.  This meant that a basic_utf8_string could be marked as noexcept(true) incorrectly as the allocator could still throw even if TINY_UTF8_NOEXCEPT was set (as T || F -> T).  Changing these to evaluate with a logical AND (i.e. T && F -> F) means that the correct noexcept specification is set.

Unit tests with a throwable and non-throwable allocator (along with TINY_UTF8_NOEXCEPT) demonstrate the correctness.

* Add missing "make test" equivalent on Windows

Fix the missing step which executes the unit tests under CI on Windows.

* Fix unit test compilation/runtime errors under Linux

Correct an issue in the CTor_TakeAnAnsiString unit test where non-ascii characters were used in a string (which meant that ansi byte encoding was different under different operating systems).  Also, adds a missing header to the mock_throwallocator.h that triggered a compilation failure under Linux (but was okay under Windows).

* Fix unit test build issue on MacOS

Building the unit tests on MacOS, using XCode, results in a compilation error when compiling code using std::equal.  This is due to the use of std::equal(it, it, it, it, predicate) which is a prototype that is only available in C++14 onwards.  For C++11, the correct prototype would be std::equal(it,it,it, predicate).  On Linux and Windows, the former works fine but this fails on MacOS.  This change corrects the unit tests so that the latter is used (which is equivalent).

This change also enforces the use of standards compliance C++11 as part of the CMake scripts.

* MacOS uses <new> not <exception> for std::bad_alloc

On MacOS, the std::bad_alloc exception definition is in the <new> header whereas, on Linux and Windows, it is defined in <exception>.  This change adds the <new> header to the Mock_ThrowAllocator header file.

* Re-add removed constructor

This change re-adds a removed constructor which supports construction using a string literal where the literal is large (and thus the SSO optimisation cannot be used).

* Fix for std::enable_if substitution failure on VS 2017 (toolset v141)

When calling the constructor (taking an utf8 char literal), VS 2017 incorrect evaluates the boolean expression and enables both SSO and non-SSO variants leading to a compilation error as the c'tors have the same prototype.  This is compounded by the Travis CI using this toolset when compiling jobs for Windows.  This change fixes this issue and permits compilation under the VS 2017 toolset.

* Add missing deps (GMock) on Focal

The dependency script was missing the gmock library under Focal which is used by the SSO unit tests.  This change adds the required packages into the Linux dependency script.

* Adjusted minor formatting of constructor in tinyutf8.h

Co-authored-by: Jakob Riedle <jakob.riedle@gmx.de>